+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - (Fix) Gracefully handle experiment declarations with 0 algorithms.
 - (Fix) #222 by @jmcolmenar: Local serializers.yml has no preference
 - (Fix) #229 by @scaverod: Solution properties are not exported to JSON
+- (Fix) Allow null customProperties in solution
 - (Deleted) Solution::isBetterThan, unused and not necessary.
 - (Deleted) Solution::getScore, Solution::recalculateScore: score recalculation should now be done in the solution validator implementation, and getScore can be any method the user wants, as long as it is declared as an Objective.
 - (Deleted) LazyMove & EagerMove: mostly unused and can be replaced by a normal Move. See SwapNeighborhood in TSP module for an example lazy implementation.